8353234: Refactor XMLSecurityPropertyManager
