+++ src/java.xml/share/classes/com/sun/org/apache/xerces/internal/parsers/DTDConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2025, Oracle and/or its affiliates. All rights reserved.
  */
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
@@ -37,7 +37,6 @@ import com.sun.org.apache.xerces.internal.impl.validation.ValidationManager;
 import com.sun.org.apache.xerces.internal.util.FeatureState;
 import com.sun.org.apache.xerces.internal.util.PropertyState;
 import com.sun.org.apache.xerces.internal.util.SymbolTable;
-import com.sun.org.apache.xerces.internal.utils.XMLSecurityPropertyManager;
 import com.sun.org.apache.xerces.internal.xni.XMLLocator;
 import com.sun.org.apache.xerces.internal.xni.XNIException;
 import com.sun.org.apache.xerces.internal.xni.grammars.XMLGrammarPool;
@@ -52,6 +51,7 @@ import javax.xml.XMLConstants;
 import javax.xml.catalog.CatalogFeatures;
 import jdk.xml.internal.JdkConstants;
 import jdk.xml.internal.JdkXmlUtils;
+import jdk.xml.internal.XMLSecurityPropertyManager;
 
 /**
  * This is the DTD-only parser configuration.  It extends the basic
@@ -89,7 +89,7 @@ import jdk.xml.internal.JdkXmlUtils;
  * @author Arnaud  Le Hors, IBM
  * @author Andy Clark, IBM
  * @author Neil Graham, IBM
- * @LastModified: May 2021
+ * @LastModified: Apr 2025
  */
 public class DTDConfiguration
     extends BasicParserConfiguration
