+++ src/java.xml/share/classes/com/sun/org/apache/xerces/internal/parsers/XIncludeAwareParserConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2023, 2025, Oracle and/or its affiliates. All rights reserved.
  */
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
@@ -24,7 +24,6 @@ import com.sun.org.apache.xerces.internal.impl.Constants;
 import com.sun.org.apache.xerces.internal.util.FeatureState;
 import com.sun.org.apache.xerces.internal.util.NamespaceSupport;
 import com.sun.org.apache.xerces.internal.util.SymbolTable;
-import com.sun.org.apache.xerces.internal.utils.XMLSecurityPropertyManager;
 import com.sun.org.apache.xerces.internal.xinclude.XIncludeHandler;
 import com.sun.org.apache.xerces.internal.xinclude.XIncludeNamespaceSupport;
 import com.sun.org.apache.xerces.internal.xni.NamespaceContext;
@@ -34,6 +33,7 @@ import com.sun.org.apache.xerces.internal.xni.parser.XMLComponentManager;
 import com.sun.org.apache.xerces.internal.xni.parser.XMLConfigurationException;
 import com.sun.org.apache.xerces.internal.xni.parser.XMLDocumentSource;
 import jdk.xml.internal.XMLSecurityManager;
+import jdk.xml.internal.XMLSecurityPropertyManager;
 
 /**
  * This class is the configuration used to parse XML 1.0 and XML 1.1 documents
@@ -41,7 +41,7 @@ import jdk.xml.internal.XMLSecurityManager;
  *
  * @author Michael Glavassevich, IBM
  *
- * @LastModified: Nov 2023
+ * @LastModified: Apr 2025
  */
 public class XIncludeAwareParserConfiguration extends XML11Configuration {
